login is not needed when there is nothing to publish

+++ ch/zbindenonline/weatherstation/publishPictures.py
@@ -21,7 +21,7 @@ class RestService:
 
     def login(self):
         logging.debug("Try to login to " + self.url + '/oauth/token')
-        logging.debug(json.dumps(self.auth))
+        # logging.debug(json.dumps(self.auth))
         try:
             loginHeaders = {'Content-Type': 'application/json'}
             response = requests.post(self.url + '/oauth/token', data=json.dumps(self.auth), headers=loginHeaders,
@@ -33,7 +33,7 @@ class RestService:
         if not response.ok:
             response.raise_for_status()
         str_response = response.content.decode('utf-8')
-        logging.debug(str_response)
+        # logging.debug(str_response)
         if str_response:
             jwtdata = json.loads(str_response)
             jwt = jwtdata['access_token']
@@ -41,8 +41,8 @@ class RestService:
             self.headers['Authorization'] = 'Bearer ' + jwt
 
     def post_picture(self, picture):
-        logging.debug('Headers:')
-        logging.debug(self.headers)
+        # logging.debug('Headers:')
+        # logging.debug(self.headers)
         filename = Path(picture).with_suffix('').name
         taken_at = datetime.datetime.strptime(filename, '%Y-%m-%d_%H%M')
         logging.debug(taken_at)
@@ -88,24 +88,27 @@ def main():
     config = read_configuration()
     configure_logging(config.loglevel)
     try:
-        service = RestService(config.pictures.picture_url, config.pictures.camera_id, config.pictures.client_id,
-                              config.pictures.client_secret,
-                              config.pictures.username, config.pictures.password)
-        pictures = get_pictures(config.pictures.picture_dir);
-        postedPictures = 0;
-        for picture in pictures:
-            logging.debug('Try to publish ' + picture)
-            try:
-                service.post_picture(picture)
-                postedPictures += 1
-                if config.pictures.delete_after_publish:
-                    logging.debug('Delete ' + picture)
-                    os.remove(picture)
-            except Exception as e:
-                logging.warning('There was en Exception in posting picture' + str(e))
+        pictures = get_pictures(config.pictures.picture_dir)
+        if len(pictures) == 0:
+            logging.info('Nothing to publish')
+        else:
+            service = RestService(config.pictures.picture_url, config.pictures.camera_id, config.pictures.client_id,
+                                  config.pictures.client_secret,
+                                  config.pictures.username, config.pictures.password)
+            postedPictures = 0
+            for picture in pictures:
+                logging.debug('Try to publish ' + picture)
+                try:
+                    service.post_picture(picture)
+                    postedPictures += 1
+                    if config.pictures.delete_after_publish:
+                        logging.debug('Delete ' + picture)
+                        os.remove(picture)
+                except Exception as e:
+                    logging.warning('There was en Exception in posting picture' + str(e))
 
-        elapsed_time = datetime.datetime.now() - start
-        logging.info('Posted ' + str(postedPictures) + ' in ' + str(elapsed_time))
+            elapsed_time = datetime.datetime.now() - start
+            logging.info('Posted ' + str(postedPictures) + ' in ' + str(elapsed_time))
     except Exception as e:
         logging.error("Error occurred: " + str(e))
